requirements.txt
requests
tqdm
mutagen
pydub
pathvalidate
pylrc
Pillow

# GUI
ttkbootstrap